Move version methods to util package
Several package (cmd, driver and cloud) could use it and to avoid import
cycle, we need to export it
